RMNE 15 - Redux (02)
On the previous commit RMNE 15 - Redux (01), there is no way for us to get the character count from the TextArea component and use it as the text for the character count text view.
To do this we will need to move the state present inside the TextArea component and place it in the Home component. This is because all the components that need access to a state have to be children of the component holding the state.

 *  move the state present inside the TextArea component and place it in the Home component
     - update Home.component.js, TextArea.component.js and TextArea\TextArea.component.android.js
     - run: npm run test:update to update Home.component.test.js.snap

Note:

By moving the state from the child component to the parent, we were able to access it in multiple children components.

Therefore, to provide access to the data that needs to be accessed by multiple components, we need to have the state in the enclosing parent component. Following this principle, if we keep on moving the state to the parent component, we will end up with the state in the topmost level component.

Redux builds on top of similar principles. It keeps a global store to which the components which need access to the data can subscribe. Additionally, it provides a mechanism by which these components can re-render whenever the data in the store changes.